\name{run_easyblup}
\alias{run_easyblup}
\title{Launch the easyblup Shiny application}
\usage{
run_easyblup(...)
}
\arguments{
\item{...}{Arguments forwarded to \code{shiny::runApp()}.}
}
\value{
This function starts the easyblup Shiny application.
}
\description{
Convenience helper that locates the packaged Shiny app assets and launches
them using \code{shiny::runApp()}.
}
